+++ cli/cli-utils/src/readProjectManifest.ts
@@ -1,18 +1,26 @@
 import * as utils from '@pnpm/read-project-manifest'
-import { type ProjectManifest } from '@pnpm/types'
+import { type SupportedArchitectures, type ProjectManifest } from '@pnpm/types'
 import { packageIsInstallable } from './packageIsInstallable'
+
+export interface ReadProjectManifestOpts {
+  engineStrict?: boolean
+  nodeVersion?: string
+  supportedArchitectures?: SupportedArchitectures
+}
+
+interface BaseReadProjectManifestResult {
+  fileName: string
+  writeProjectManifest: (manifest: ProjectManifest, force?: boolean) => Promise<void>
+}
+
+export interface ReadProjectManifestResult extends BaseReadProjectManifestResult {
+  manifest: ProjectManifest
+}
 
 export async function readProjectManifest (
   projectDir: string,
-  opts: {
-    engineStrict?: boolean
-    nodeVersion?: string
-  }
-): Promise<{
-    fileName: string
-    manifest: ProjectManifest
-    writeProjectManifest: (manifest: ProjectManifest, force?: boolean) => Promise<void>
-  }> {
+  opts: ReadProjectManifestOpts = {}
+): Promise<ReadProjectManifestResult> {
   const { fileName, manifest, writeProjectManifest } = await utils.readProjectManifest(projectDir)
   packageIsInstallable(projectDir, manifest as any, opts) // eslint-disable-line @typescript-eslint/no-explicit-any
   return { fileName, manifest, writeProjectManifest }
@@ -20,27 +28,21 @@ export async function readProjectManifest (
 
 export async function readProjectManifestOnly (
   projectDir: string,
-  opts: {
-    engineStrict?: boolean
-    nodeVersion?: string
-  } = {}
+  opts: ReadProjectManifestOpts = {}
 ): Promise<ProjectManifest> {
   const manifest = await utils.readProjectManifestOnly(projectDir)
   packageIsInstallable(projectDir, manifest as any, opts) // eslint-disable-line @typescript-eslint/no-explicit-any
   return manifest
 }
 
+export interface TryReadProjectManifestResult extends BaseReadProjectManifestResult {
+  manifest: ProjectManifest | null
+}
+
 export async function tryReadProjectManifest (
   projectDir: string,
-  opts: {
-    engineStrict?: boolean
-    nodeVersion?: string
-  }
-): Promise<{
-    fileName: string
-    manifest: ProjectManifest | null
-    writeProjectManifest: (manifest: ProjectManifest, force?: boolean) => Promise<void>
-  }> {
+  opts: ReadProjectManifestOpts
+): Promise<TryReadProjectManifestResult> {
   const { fileName, manifest, writeProjectManifest } = await utils.tryReadProjectManifest(projectDir)
   if (manifest == null) return { fileName, manifest, writeProjectManifest }
   packageIsInstallable(projectDir, manifest as any, opts) // eslint-disable-line @typescript-eslint/no-explicit-any