+++ config/package-is-installable/src/checkPlatform.ts
@@ -1,4 +1,5 @@
 import { PnpmError } from '@pnpm/error'
+import { type SupportedArchitectures } from '@pnpm/types'
 import { familySync as getLibcFamilySync } from 'detect-libc'
 
 const currentLibc = getLibcFamilySync() ?? 'unknown'
@@ -16,19 +17,26 @@ export class UnsupportedPlatformError extends PnpmError {
 
 export function checkPlatform (
   packageId: string,
-  wantedPlatform: WantedPlatform
+  wantedPlatform: WantedPlatform,
+  supportedArchitectures?: SupportedArchitectures
 ) {
+  const current = {
+    os: dedupeCurrent(process.platform, supportedArchitectures?.os ?? ['current']),
+    cpu: dedupeCurrent(process.arch, supportedArchitectures?.cpu ?? ['current']),
+    libc: dedupeCurrent(currentLibc, supportedArchitectures?.libc ?? ['current']),
+  }
+
   const { platform, arch } = process
   let osOk = true; let cpuOk = true; let libcOk = true
 
   if (wantedPlatform.os) {
-    osOk = checkList(platform, wantedPlatform.os)
+    osOk = checkList(current.os, wantedPlatform.os)
   }
   if (wantedPlatform.cpu) {
-    cpuOk = checkList(arch, wantedPlatform.cpu)
+    cpuOk = checkList(current.cpu, wantedPlatform.cpu)
   }
   if (wantedPlatform.libc && currentLibc !== 'unknown') {
-    libcOk = checkList(currentLibc, wantedPlatform.libc)
+    libcOk = checkList(current.libc, wantedPlatform.libc)
   }
 
   if (!osOk || !cpuOk || !libcOk) {
@@ -45,25 +53,35 @@ export interface Platform {
 
 export type WantedPlatform = Partial<Platform>
 
-function checkList (value: string, list: string | string[]) {
-  let tmp; let match = false; let blc = 0
+function checkList (value: string | string[], list: string | string[]): boolean {
+  let tmp
+  let match = false
+  let blc = 0
+
   if (typeof list === 'string') {
     list = [list]
   }
   if (list.length === 1 && list[0] === 'any') {
     return true
   }
-  for (let i = 0; i < list.length; ++i) {
-    tmp = list[i]
-    if (tmp[0] === '!') {
-      tmp = tmp.slice(1)
-      if (tmp === value) {
-        return false
+  const values = Array.isArray(value) ? value : [value]
+  for (const value of values) {
+    for (let i = 0; i < list.length; ++i) {
+      tmp = list[i]
+      if (tmp[0] === '!') {
+        tmp = tmp.slice(1)
+        if (tmp === value) {
+          return false
+        }
+        ++blc
+      } else {
+        match = match || tmp === value
       }
-      ++blc
-    } else {
-      match = match || tmp === value
     }
   }
   return match || blc === list.length
+}
+
+function dedupeCurrent (current: string, supported: string[]) {
+  return supported.map((supported) => supported === 'current' ? current : supported)
 }