+++ config/package-is-installable/test/checkPlatform.ts
@@ -79,3 +79,45 @@ test('os wrong (negation)', () => {
 test('nothing wrong (negation)', () => {
   expect(checkPlatform(packageId, { cpu: '!enten-cpu', os: '!enten-os', libc: '!enten-libc' })).toBe(null)
 })
+
+test('override OS', () => {
+  expect(checkPlatform(packageId, { cpu: 'any', os: 'win32', libc: 'any' }, {
+    os: ['win32'],
+    cpu: ['current'],
+    libc: ['current'],
+  })).toBe(null)
+})
+
+test('accept another CPU', () => {
+  expect(checkPlatform(packageId, { cpu: 'x64', os: 'any', libc: 'any' }, {
+    os: ['current'],
+    cpu: ['current', 'x64'],
+    libc: ['current'],
+  })).toBe(null)
+})
+
+test('fail when CPU is different', () => {
+  const err = checkPlatform(packageId, { cpu: 'x64', os: 'any', libc: 'any' }, {
+    os: ['current'],
+    cpu: ['arm64'],
+    libc: ['current'],
+  })
+  expect(err).toBeTruthy()
+  expect(err?.code).toBe('ERR_PNPM_UNSUPPORTED_PLATFORM')
+})
+
+test('override libc', () => {
+  expect(checkPlatform(packageId, { cpu: 'any', os: 'any', libc: 'glibc' }, {
+    os: ['current'],
+    cpu: ['current'],
+    libc: ['glibc'],
+  })).toBe(null)
+})
+
+test('accept another libc', () => {
+  expect(checkPlatform(packageId, { cpu: 'any', os: 'any', libc: 'glibc' }, {
+    os: ['current'],
+    cpu: ['current'],
+    libc: ['current', 'glibc'],
+  })).toBe(null)
+})