+++ patching/plugin-commands-patching/src/getPatchedDependency.ts
@@ -1,16 +1,17 @@
 import path from 'path'
 import { parseWantedDependency, type ParseWantedDependencyResult } from '@pnpm/parse-wanted-dependency'
 import { prompt } from 'enquirer'
-import { readCurrentLockfile } from '@pnpm/lockfile-file'
+import { readCurrentLockfile, type TarballResolution } from '@pnpm/lockfile-file'
 import { nameVerFromPkgSnapshot } from '@pnpm/lockfile-utils'
 import { PnpmError } from '@pnpm/error'
 import { WANTED_LOCKFILE } from '@pnpm/constants'
 import { readModulesManifest } from '@pnpm/modules-yaml'
+import { isGitHostedPkgUrl } from '@pnpm/pick-fetcher'
 import realpathMissing from 'realpath-missing'
 import semver from 'semver'
 import { type Config } from '@pnpm/config'
 
-type GetPatchedDependencyOptions = {
+export type GetPatchedDependencyOptions = {
   lockfileDir: string
 } & Pick<Config, 'virtualStoreDir' | 'modulesDir'>
 
@@ -22,7 +23,7 @@ export async function getPatchedDependency (rawDependency: string, opts: GetPatc
   if (!preferredVersions.length) {
     throw new PnpmError(
       'PATCH_VERSION_NOT_FOUND',
-      `Can not find ${rawDependency} in project ${opts.lockfileDir}, ${versions.length ? `you can specify currently installed version: ${versions.join(', ')}.` : `did you forget to install ${rawDependency}?`}`
+      `Can not find ${rawDependency} in project ${opts.lockfileDir}, ${versions.length ? `you can specify currently installed version: ${versions.map(({ version }) => version).join(', ')}.` : `did you forget to install ${rawDependency}?`}`
     )
   }
 
@@ -34,17 +35,26 @@ export async function getPatchedDependency (rawDependency: string, opts: GetPatc
       type: 'select',
       name: 'version',
       message: 'Choose which version to patch',
-      choices: versions,
+      choices: preferredVersions.map(preferred => ({
+        name: preferred.version,
+        message: preferred.version,
+        value: preferred.gitTarballUrl ?? preferred.version,
+        hint: preferred.gitTarballUrl ? 'Git Hosted' : undefined,
+      })),
+      result (selected) {
+        const selectedVersion = preferredVersions.find(preferred => preferred.version === selected)!
+        return selectedVersion.gitTarballUrl ?? selected
+      },
     })
     dep.pref = version
   } else {
-    dep.pref = preferredVersions[0]
+    const preferred = preferredVersions[0]
+    dep.pref = preferred.gitTarballUrl ?? preferred.version
   }
-
   return dep
 }
 
-async function getVersionsFromLockfile (dep: ParseWantedDependencyResult, opts: GetPatchedDependencyOptions) {
+export async function getVersionsFromLockfile (dep: ParseWantedDependencyResult, opts: GetPatchedDependencyOptions) {
   const modulesDir = await realpathMissing(path.join(opts.lockfileDir, opts.modulesDir ?? 'node_modules'))
   const modules = await readModulesManifest(modulesDir)
   const lockfile = (modules?.virtualStoreDir && await readCurrentLockfile(modules.virtualStoreDir, {
@@ -61,12 +71,17 @@ async function getVersionsFromLockfile (dep: ParseWantedDependencyResult, opts: 
   const pkgName = dep.alias && dep.pref ? dep.alias : (dep.pref ?? dep.alias)
 
   const versions = Object.entries(lockfile.packages ?? {})
-    .map(([depPath, pkgSnapshot]) => nameVerFromPkgSnapshot(depPath, pkgSnapshot))
+    .map(([depPath, pkgSnapshot]) => {
+      const tarball = (pkgSnapshot.resolution as TarballResolution)?.tarball ?? ''
+      return {
+        ...nameVerFromPkgSnapshot(depPath, pkgSnapshot),
+        gitTarballUrl: isGitHostedPkgUrl(tarball) ? tarball : undefined,
+      }
+    })
     .filter(({ name }) => name === pkgName)
-    .map(({ version }) => version)
 
   return {
     versions,
-    preferredVersions: versions.filter(version => dep.alias && dep.pref ? semver.satisfies(version, dep.pref) : true),
+    preferredVersions: versions.filter(({ version }) => dep.alias && dep.pref ? semver.satisfies(version, dep.pref) : true),
   }
 }