+++ pkg-manager/plugin-commands-installation/test/fetch.ts
@@ -28,6 +28,7 @@ const DEFAULT_OPTIONS = {
   registries: {
     default: REGISTRY_URL,
   },
+  rootProjectManifestDir: '',
   sort: true,
   userConfig: {},
   workspaceConcurrency: 1,