+++ pkg-manager/plugin-commands-installation/test/update/interactive.ts
@@ -37,6 +37,7 @@ const DEFAULT_OPTIONS = {
   registries: {
     default: REGISTRY_URL,
   },
+  rootProjectManifestDir: '',
   sort: true,
   userConfig: {},
   workspaceConcurrency: 1,