+++ reviewing/plugin-commands-outdated/test/utils/index.ts
@@ -41,6 +41,7 @@ export const DEFAULT_OPTS = {
   rawLocalConfig: {},
   registries: { default: REGISTRY },
   registry: REGISTRY,
+  rootProjectManifestDir: '',
   sort: true,
   storeDir: '../store',
   strictSsl: false,