+++ __utils__/assert-project/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pnpm/assert-project
 
+## 2.3.40
+
+### Patch Changes
+
+- Updated dependencies [43ce9e4a6]
+  - @pnpm/types@9.4.0
+  - @pnpm/lockfile-types@5.1.3
+  - @pnpm/modules-yaml@12.1.4
+  - @pnpm/assert-store@1.0.75
+
 ## 2.3.39
 
 ### Patch Changes