+++ dedupe/check/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm/dedupe.check
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [43ce9e4a6]
+  - @pnpm/types@9.4.0
+  - @pnpm/lockfile-types@5.1.3
+
 ## 1.0.4
 
 ### Patch Changes