+++ env/node.fetcher/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pnpm/node.fetcher
 
+## 3.0.28
+
+### Patch Changes
+
+- @pnpm/create-cafs-store@6.0.6
+- @pnpm/fetcher-base@15.0.3
+- @pnpm/tarball-fetcher@18.0.8
+- @pnpm/pick-fetcher@2.0.1
+
 ## 3.0.27
 
 ### Patch Changes