+++ fetching/fetcher-base/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm/fetcher-base
 
+## 15.0.3
+
+### Patch Changes
+
+- Updated dependencies [43ce9e4a6]
+  - @pnpm/types@9.4.0
+  - @pnpm/resolver-base@10.0.4
+
 ## 15.0.2
 
 ### Patch Changes