+++ fetching/git-fetcher/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pnpm/git-fetcher
 
+## 12.0.8
+
+### Patch Changes
+
+- Updated dependencies [6390033cd]
+  - @pnpm/worker@0.3.6
+  - @pnpm/prepare-package@5.0.16
+  - @pnpm/fetcher-base@15.0.3
+
 ## 12.0.7
 
 ### Patch Changes