+++ hooks/read-package-hook/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pnpm/hooks.read-package-hook
 
+## 3.0.8
+
+### Patch Changes
+
+- Updated dependencies [43ce9e4a6]
+  - @pnpm/types@9.4.0
+
 ## 3.0.7
 
 ### Patch Changes