+++ lockfile/lockfile-walker/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pnpm/lockfile-walker
 
+## 7.0.6
+
+### Patch Changes
+
+- Updated dependencies [43ce9e4a6]
+  - @pnpm/types@9.4.0
+  - @pnpm/lockfile-types@5.1.3
+  - @pnpm/dependency-path@2.1.5
+
 ## 7.0.5
 
 ### Patch Changes