+++ resolving/tarball-resolver/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pnpm/tarball-resolver
 
+## 8.0.4
+
+### Patch Changes
+
+- @pnpm/resolver-base@10.0.4
+
 ## 8.0.3
 
 ### Patch Changes