+++ store/cafs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm/store.cafs
 
+## 2.0.7
+
+### Patch Changes
+
+- Updated dependencies [43ce9e4a6]
+  - @pnpm/store-controller-types@17.1.0
+  - @pnpm/fetcher-base@15.0.3
+
 ## 2.0.6
 
 ### Patch Changes